Add support for providing data file to new images.
When booting a new node we need to pass along the application code
and any other resources required for the site. This commit provides the
feature for the KVM images, using the OVF ISO image that gets attached
to the VM. Amazon EC2 will use S3 to pass this data.