Add deployment files for deploying to Ubuntu.
Node app is deployed using a cloud-init userdata script to configure
a stock Ubuntu cloud image. NodeJS is installed from the chris-lea ppa
and runs on port 8000. Nginx listens on port 80 and proxies requests
to nodejs.

Node.js is currently running as root - this needs to be fixed eventually
for security reasons.